Добавлена регистрация, идентификация и аутентификация

+++ src/main/java/ru/bahusdivus/bhope/controllers/PostsController.java
@@ -1,25 +1,37 @@
 package ru.bahusdivus.bhope.controllers;
 
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import ru.bahusdivus.bhope.dto.CommentDto;
 import ru.bahusdivus.bhope.dto.PostWithCommentsDto;
 import ru.bahusdivus.bhope.dto.UserDto;
 import ru.bahusdivus.bhope.services.CommentsService;
+import ru.bahusdivus.bhope.services.UserService;
 
 @Controller
 public class PostsController {
 
     private final CommentsService commentsService;
+    private final UserService userService;
 
-    public PostsController(CommentsService commentsService) {
+
+    public PostsController(CommentsService commentsService, UserService userService) {
         this.commentsService = commentsService;
+        this.userService = userService;
     }
 
     @RequestMapping("/post/{postId}")
-    public String getPage(@PathVariable("postId") long postId, Model model) {
+    public String getPage(@PathVariable("postId") long postId,
+                          @AuthenticationPrincipal UserDetails userDetails,
+                          Model model) {
         PostWithCommentsDto post = commentsService.getPost(postId);
+        model.addAttribute("login", userDetails != null ? userDetails.getUsername() : null);
         model.addAttribute("title", "Заголовок");
         model.addAttribute("postId", postId);
         model.addAttribute("content", post.getContent());
@@ -28,29 +40,51 @@ public class PostsController {
     }
 
     @RequestMapping("/post/{postId}/comment")
-    public String getNewCommentForm(@PathVariable("postId") long postId, Model model) {
-        return getNewCommentFormWithParent(postId, 0, model);
+    public String getNewCommentForm(@PathVariable("postId") long postId,
+                                    @AuthenticationPrincipal UserDetails userDetails,
+                                    Model model) throws IllegalAccessException {
+        return getNewCommentFormWithParent(postId, 0, userDetails, model);
     }
 
     @RequestMapping("/post/{postId}/comment/{id}")
-    public String getEditCommentForm(@PathVariable("postId") long postId, @PathVariable("id") long id, Model model) {
+    public String getEditCommentForm(@PathVariable("postId") long postId,
+                                     @PathVariable("id") long id,
+                                     @AuthenticationPrincipal UserDetails userDetails,
+                                     Model model) throws IllegalAccessException {
+        if (userDetails == null) {
+            throw new IllegalAccessException("нужно зарегистрироваться");
+        }
         CommentDto comment = commentsService.getComment(id);
         model.addAttribute("comment", comment);
         return "commentPage";
     }
 
     @RequestMapping("/post/{postId}/comment/{id}/comment")
-    public String getNewCommentFormWithParent(@PathVariable("postId") long postId, @PathVariable("id") long id, Model model) {
+    public String getNewCommentFormWithParent(@PathVariable("postId") long postId,
+                                              @PathVariable("id") long id,
+                                              @AuthenticationPrincipal UserDetails userDetails,
+                                              Model model) throws IllegalAccessException {
+
+        if (userDetails == null) {
+            throw new IllegalAccessException("нужно зарегистрироваться");
+        }
         CommentDto commentDto = new CommentDto();
         commentDto.setPost(postId);
         commentDto.setParent(id);
-        commentDto.setUser(new UserDto(1));
+        long userId = userService.findByLogin(userDetails.getUsername()).getId();
+        commentDto.setUser(new UserDto(userId));
+        model.addAttribute("login", userDetails.getUsername());
         model.addAttribute("comment", commentDto);
         return "commentPage";
     }
 
     @RequestMapping("/post/{postId}/comment/{id}/delete")
-    public String deleteComment(@PathVariable("postId") long postId, @PathVariable("id") long id) {
+    public String deleteComment(@PathVariable("postId") long postId,
+                                @PathVariable("id") long id,
+                                @AuthenticationPrincipal UserDetails userDetails) throws IllegalAccessException {
+        if (userDetails == null) {
+            throw new IllegalAccessException("нужно зарегистрироваться");
+        }
         commentsService.deleteComment(id);
         return "redirect:/post/" + postId;
     }
